K10-3548-2 Refactor to keep copied packages seperate
