+++ pkg/blockstorage/zone/zone.go
@@ -221,3 +221,27 @@ func NodeZonesAndRegion(ctx context.Context, cli kubernetes.Interface) (map[stri
 	}
 	return zoneSet, region[0], nil
 }
+
+func sanitizeAvailableZones(availableZones map[string]struct{}, validZoneNames []string) map[string]struct{} {
+	sanitizedZones := map[string]struct{}{}
+	for zone := range availableZones {
+		if isZoneNameValid(zone, validZoneNames) {
+			sanitizedZones[zone] = struct{}{}
+		} else {
+			closestMatch := levenshteinMatch(zone, validZoneNames)
+			log.Debug().Print("Exact match not found for available zone, using closest match",
+				field.M{"availableZone": zone, "closestMatch": closestMatch})
+			sanitizedZones[closestMatch] = struct{}{}
+		}
+	}
+	return sanitizedZones
+}
+
+func isZoneNameValid(zone string, validZones []string) bool {
+	for _, z := range validZones {
+		if zone == z {
+			return true
+		}
+	}
+	return false
+}
